+++ src/main/java/com/feeyo/redis/engine/codec/RedisResponseDecoderV4.java
@@ -14,9 +14,9 @@ public class RedisResponseDecoderV4 {
 	private byte[] _buffer;
 	private int _offset;
 	
-	private List<RedisResponseV3> responses = null;
+	private List<RedisResponse> responses = null;
 	
-	public List<RedisResponseV3> decode(byte[] buffer) {
+	public List<RedisResponse> decode(byte[] buffer) {
 
 		append(buffer);
 
@@ -24,7 +24,7 @@ public class RedisResponseDecoderV4 {
 			if ( responses != null )
 				responses.clear();
 			else 
-				responses = new ArrayList<RedisResponseV3>(2);
+				responses = new ArrayList<RedisResponse>(2);
 			
 			for(;;) {
 				
@@ -40,7 +40,7 @@ public class RedisResponseDecoderV4 {
 		          case '-':					// 错误, 表示错误的状态信息, "-" 后就是具体信息
 		          case ':':					// 整数, 以 ":" 开头, 返回
 		          case '$':					// 批量字符串, 以 "$" 开头,表示下一行的字符串长度,具体字符串在下一行中,字符串最大能达到512MB  
-		            RedisResponseV3 resp = parseResponse(type);
+		            RedisResponse resp = parseResponse(type);
 		            if ( resp != null ) {
 		            	responses.add( resp );
 		            }
@@ -64,7 +64,7 @@ public class RedisResponseDecoderV4 {
 		
 	}
 	
-	private RedisResponseV3 parseResponse(byte type) {
+	private RedisResponse parseResponse(byte type) {
 
 		int start, end, len, offset;
 		
@@ -83,7 +83,7 @@ public class RedisResponseDecoderV4 {
 
 			// 长度
 			len = end - start;
-			return new RedisResponseV3(type, getBytes(start, len));
+			return new RedisResponse(type, getBytes(start, len));
 			
 		 } else if (type == '$') {
 			 
@@ -95,7 +95,7 @@ public class RedisResponseDecoderV4 {
 		    	start = offset - 1;
 		    	end = _offset;
 		    	len = end - start;
-		        return new RedisResponseV3(type, getBytes(start, len));  // 此处不减
+		        return new RedisResponse(type, getBytes(start, len));  // 此处不减
 		      }
 
 		      end = _offset + packetSize + 2;	// offset + data + \r\n
@@ -108,7 +108,7 @@ public class RedisResponseDecoderV4 {
 		      
 		      start = offset - 1;
 		      len = end - start;
-		      return new RedisResponseV3(type, getBytes(start, len));
+		      return new RedisResponse(type, getBytes(start, len));
 			 
 		 } else if (type == '*') {
 			 
@@ -120,7 +120,7 @@ public class RedisResponseDecoderV4 {
 		    	start = offset -1;
 		    	end = _offset;
 		    	len = end - start;
-		        return new RedisResponseV3(type, getBytes(start, len));  // 此处不减
+		        return new RedisResponse(type, getBytes(start, len));  // 此处不减
 		      }
 
 		      if (packetSize > bytesRemaining()) {
@@ -129,14 +129,14 @@ public class RedisResponseDecoderV4 {
 		      }
 
 		      // 此处多增加一长度，用于存储 *packetSize\r\n
-		      RedisResponseV3 response = new RedisResponseV3(type, packetSize + 1);
+		      RedisResponse response = new RedisResponse(type, packetSize + 1);
 		      start = offset -1;
 		      end = _offset;
 		      len = end - start;
-		      response.set(0, new RedisResponseV3(type, getBytes(start, len)));
+		      response.set(0, new RedisResponse(type, getBytes(start, len)));
 
 		      byte ntype;
-		      RedisResponseV3 res;
+		      RedisResponse res;
 		      for (int i = 1; i <= packetSize; i++) {
 		        if (_offset + 1 >= _buffer.length ) {
 		          throw new IndexOutOfBoundsException("Wait for more data.");
@@ -274,7 +274,7 @@ public class RedisResponseDecoderV4 {
 		RedisResponseDecoderV4 decoder = new RedisResponseDecoderV4();
 		//List<RedisResponseV3> resps = decoder.decode(buffer);
 		
-		List<RedisResponseV3> resps = decoder.decode(buffer);
+		List<RedisResponse> resps = decoder.decode(buffer);
 		System.out.println( resps );
 		resps = decoder.decode(buffer2);
 //		System.out.println( resps );
